requirements.txt
discord.py
alpaca-trade-api
python-dotenv
flask
pandas
scikit-learn
joblib
schedule
twilio
